Adding requirements.dev.txt to Dockerfile
This way we can ensure that:
	If we use docker-compose build
	command to build docker images
	we are overriding DEV argument
	as true and installing flake8 (linting tool)
	for only DEV environment.

Glossary:

WORKDIR /app - ensures that furather RUN COPY commands will
address files right in the /app/ directory. We won't need to
add it always in the beginning with this.

ENV PATH - PATH variable is used to find "executable". With this
we ensure that we are using executable versions that are present
inside the virtual environment of the container that we are raising.
For example, if we use pythong in our docker container, then we are
sure that we are using pythong version specific for our virtual environment
inside our docker container.

fi && - is a special syntax of Dockerfile to end if statement.

python:3.9-alpine3.13 - is one of the lightest pythong images, just perfect for
our cast of simple REST API

ENV PYTHONBUFFERED 1 - this variable affects behavior of Python's standard I/O streams.
With this variable as 1, the performance improves.